Issue #229: Extracting an interface from AnalyzerFutureResult. Introducing ToStringFutureTextRenderer.
