+++ api/src/main/java/org/datacleaner/api/AnalyzerResultFuture.java
@@ -19,17 +19,8 @@
  */
 package org.datacleaner.api;
 
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-
 import org.apache.metamodel.util.HasName;
 import org.apache.metamodel.util.Ref;
-import org.apache.metamodel.util.SharedExecutorService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Represents an {@link AnalyzerResult} that is still being produced.
@@ -43,7 +34,7 @@ import org.slf4j.LoggerFactory;
  * @param <R>
  *            the wrapped {@link AnalyzerResult} type.
  */
-public class AnalyzerResultFuture<R extends AnalyzerResult> implements AnalyzerResult, HasName, Ref<R> {
+public interface AnalyzerResultFuture<R extends AnalyzerResult> extends AnalyzerResult, HasName, Ref<R> {
 
     /**
      * Listener interface for objects that want to be notified when the wrapped
@@ -58,121 +49,28 @@ public class AnalyzerResultFuture<R extends AnalyzerResult> implements AnalyzerR
         public void onError(RuntimeException error);
     }
 
-    private static final Logger logger = LoggerFactory.getLogger(AnalyzerResultFuture.class);
-
-    private static final long serialVersionUID = 1L;
-
-    private transient final CountDownLatch _countDownLatch;
-    private transient List<Listener<? super R>> _listeners;
-
-    private final String _name;
-    private R _result;
-    private RuntimeException _error;
-
     /**
-     * Constructs an {@link AnalyzerResultFuture}
+     * Gets the name of the {@link AnalyzerResult} that is being waited on. This
+     * is useful for presenting to the user what he is waiting for.
      * 
-     * @param name
-     *            a name/label to use for presenting and distinguishing this
-     *            result from others.
-     * @param resultRef
-     *            a reference for the result being processed.
+     * @return the name of the result being waited for
      */
-    public AnalyzerResultFuture(String name, final Ref<? extends R> resultRef) {
-        _name = name;
-        _countDownLatch = new CountDownLatch(1);
-        _result = null;
-        _error = null;
-
-        SharedExecutorService.get().submit(new Runnable() {
-
-            @Override
-            public void run() {
-                try {
-                    _result = resultRef.get();
-                    onSuccess();
-                } catch (RuntimeException e) {
-                    _error = e;
-                    onError();
-                } finally {
-                    _countDownLatch.countDown();
-                }
-            }
-        });
-    }
+    @Override
+    public String getName();
 
     /**
      * Adds a {@link Listener} to this {@link AnalyzerResultFuture}.
      * 
      * @param listener
      */
-    public synchronized void addListener(Listener<? super R> listener) {
-        // it might be we add a listener AFTER the result is actually produced,
-        // in which case we simply inform the listener immediately.
-        if (isReady()) {
-            if (_error != null) {
-                listener.onError(_error);
-            } else {
-                listener.onSuccess(_result);
-            }
-            return;
-        }
-
-        if (_listeners == null) {
-            _listeners = new LinkedList<>();
-        }
-        _listeners.add(listener);
-    }
+    public void addListener(Listener<? super R> listener);
 
     /**
      * Removes a {@link Listener} from this {@link AnalyzerResultFuture}.
      * 
      * @param listener
      */
-    public synchronized void removeListener(Listener<R> listener) {
-        if (_listeners == null) {
-            return;
-        }
-        _listeners.remove(listener);
-    }
-
-    private synchronized void onSuccess() {
-        if (_listeners == null) {
-            return;
-        }
-        try {
-            for (final Listener<? super R> listener : _listeners) {
-                try {
-                    listener.onSuccess(_result);
-                } catch (Exception e) {
-                    logger.warn("Unexpected exception while informing listener of success: {}", listener, e);
-                }
-            }
-        } catch (Exception e) {
-            logger.warn("Unexpected exception while iterating listeners on success", e);
-        } finally {
-            _listeners = null;
-        }
-    }
-
-    private synchronized void onError() {
-        if (_listeners == null) {
-            return;
-        }
-        try {
-            for (final Listener<? super R> listener : _listeners) {
-                try {
-                    listener.onError(_error);
-                } catch (Exception e) {
-                    logger.warn("Unexpected exception while informing listener on error: {}", listener, e);
-                }
-            }
-        } catch (Exception e) {
-            logger.warn("Unexpected exception while iterating listeners on error", e);
-        } finally {
-            _listeners = null;
-        }
-    }
+    public void removeListener(Listener<R> listener);
 
     /**
      * Determines if the wrapped {@link AnalyzerResult} is ready or if
@@ -183,50 +81,14 @@ public class AnalyzerResultFuture<R extends AnalyzerResult> implements AnalyzerR
      * @return true if the wrapped {@link AnalyzerResult} is ready or false if
      *         it is not.
      */
-    public boolean isReady() {
-        if (_countDownLatch == null) {
-            return true;
-        }
-
-        return _countDownLatch.getCount() == 0;
-    }
-
-    @Override
-    public R get() {
-        if (_countDownLatch != null) {
-            try {
-                _countDownLatch.await();
-            } catch (InterruptedException e) {
-                // do nothing
-            }
-        }
-        if (_error != null) {
-            throw _error;
-        }
-        return _result;
-    }
-
-    @Override
-    public String toString() {
-        return "AnalyzerResultFuture[" + _name + "]";
-    }
-
-    @Override
-    public String getName() {
-        return _name;
-    }
+    public boolean isReady();
 
     /**
-     * Method invoked when serialization takes place. Makes sure that we await
-     * the loading of the result reference before writing any data.
+     * Gets (and awaits the processing in a blocking fashion if necesary) the
+     * wrapped {@link AnalyzerResult}.
      * 
-     * @param out
-     * @throws IOException
+     * @return the wrapped {@link AnalyzerResult}
      */
-    private void writeObject(ObjectOutputStream out) throws IOException {
-        logger.info("Serialization requested, awaiting reference to load.");
-        get();
-        out.defaultWriteObject();
-        logger.info("Serialization finished!");
-    }
+    @Override
+    public R get();
 }
