+++ testware/src/main/java/org/datacleaner/result/MockAnalyzerFutureResult.java
@@ -30,5 +30,10 @@ public class MockAnalyzerFutureResult implements AnalyzerResult {
     public String getMockMessage() {
         return "MockAnalyzerResult available";
     }
+    
+    @Override
+    public String toString() {
+        return getMockMessage();
+    }
 
 }
